example commit tampering with the library
This commit creates an unofficial modification of the 'requests' PyPI
packages for the purposes of demonstrating SLSA and tags it as a (fake)
2.31.666 release.

Note that this commit only exists on the MarkLodato/requests fork and
not the official psf/requests fork. I do not have write access to the
latter. Because MarkLodato/requests is not in the SLSA policy for
`requests`, builds from this repository should not be accepted.

See https://github.com/MarkLodato/slsa-playground for more details.